Moved some classes around as the names were too long
This is because even though its in the same module its being named the module again which is honestly redundant
