+++ app/src/main/java/com/itachi1706/cheesecakeutilities/Modules/gpaCalculator/fragment/SemesterListFragment.kt
@@ -10,13 +10,16 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.DefaultItemAnimator
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.google.android.material.snackbar.Snackbar
 import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.ValueEventListener
-import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.GpaCalculatorMainActivity
-import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.interfaces.GpaCalcCallback
+import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.GpaCalcFirebaseUtils
+import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.MainViewActivity
 import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.interfaces.StateSwitchListener
 import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.objects.GpaInstitution
+import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.objects.GpaScoring
+import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.objects.GpaSemester
 import com.itachi1706.cheesecakeutilities.R
 import com.itachi1706.cheesecakeutilities.RecyclerAdapters.DualLineStringRecyclerAdapter
 import com.itachi1706.cheesecakeutilities.Util.FirebaseUtils
@@ -24,19 +27,25 @@ import com.itachi1706.cheesecakeutilities.Util.LogHelper
 import com.itachi1706.cheesecakeutilities.objects.DualLineString
 
 /**
- * Institution List View
+ * Semester List View
  */
-class GpaCalcInstitutionListFragment : Fragment() {
+class SemesterListFragment : Fragment() {
 
     private var callback: StateSwitchListener? = null
-    private val state = GpaCalculatorMainActivity.STATE_INSTITUTION
+    private val state = MainViewActivity.STATE_SEMESTER
 
-    private val institutions: ArrayList<GpaInstitution> = ArrayList()
+    private val semesters: ArrayList<GpaSemester> = ArrayList()
+
     private lateinit var adapter: DualLineStringRecyclerAdapter
+    private var selectedInstitutionString: String? = null
+    private var selectedInstitutionType: String? = null
+
+    private var selectedInstitution: GpaInstitution? = null
+    private var scoreObject: GpaScoring? = null
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
-        if (context is GpaCalculatorMainActivity) {
+        if (context is MainViewActivity) {
             callback = context
         }
     }
@@ -45,6 +54,15 @@ class GpaCalcInstitutionListFragment : Fragment() {
                               savedInstanceState: Bundle?): View? {
         val v = inflater.inflate(R.layout.fragment_recycler_view, container, false)
         val recyclerView = v.findViewById<RecyclerView>(R.id.main_menu_recycler_view)
+
+        selectedInstitutionString = arguments?.getString("selection")
+        if (selectedInstitutionString == null) {
+            LogHelper.e(TAG, "Institution not selected!")
+            Snackbar.make(v, "An error has occurred. (Institution not found)", Snackbar.LENGTH_LONG).show()
+            callback?.goBack()
+            return v
+        }
+        selectedInstitutionType = arguments?.getString("type")
 
         recyclerView.setHasFixedSize(true)
         val linearLayoutManager = LinearLayoutManager(context)
@@ -61,42 +79,45 @@ class GpaCalcInstitutionListFragment : Fragment() {
         adapter.setOnClickListener { view ->
             val viewHolder = view.tag as DualLineStringRecyclerAdapter.StringViewHolder
             val pos = viewHolder.adapterPosition
-            val instituteSelected = institutions[pos]
-            callback?.selectInstitute(instituteSelected)
+            val semesterSelected = semesters[pos]
+            // TODO: Switch fragment with the selected semester
+            Snackbar.make(view, "Unimplemented. Selected: ${semesterSelected.name}", Snackbar.LENGTH_LONG).show()
         }
-        return v
-    }
 
-    override fun onResume() {
-        super.onResume()
-        callback?.updateActionBar(null, null)
+        // Get institution name to update title
+        selectedInstitution = callback?.getInstitution()
+        updateActionBar()
+        return v
     }
 
     private var listener: ValueEventListener? = null
     override fun onStart() {
         super.onStart()
+        if (selectedInstitutionString == null) return // Don't do anything, an error had occurred already
         if (listener != null) {
             FirebaseUtils.removeListener(listener!!)
             listener = null
             LogHelper.e(TAG, "Firebase DB Listeners exists when it should not have, terminating it forcibly")
         }
 
-        LogHelper.i(TAG, "Registering Institution Firebase DB Listeners")
-        listener = callback?.getUserData()?.addValueEventListener(object: ValueEventListener{
+        scoreObject = callback?.getScoreMap()!![selectedInstitutionType]
+        updateActionBar()
+        LogHelper.i(TAG, "Registering Semester Firebase DB Listeners")
+        listener = callback?.getUserData()?.child(selectedInstitutionString!!)?.child(GpaCalcFirebaseUtils.FB_REC_SEMESTER)?.addValueEventListener(object: ValueEventListener{
             override fun onCancelled(p0: DatabaseError) {
-                LogHelper.w(TAG, "loadInstitutionsList:onCancelled", p0.toException())
+                LogHelper.w(TAG, "loadSemestersList:onCancelled", p0.toException())
             }
 
             override fun onDataChange(dataSnapshot: DataSnapshot) {
-                LogHelper.i(TAG, "Processing updated institutions...")
-                institutions.clear()
+                LogHelper.i(TAG, "Processing updated semesters...")
+                semesters.clear()
                 if (!dataSnapshot.hasChildren()) return
                 dataSnapshot.children.forEach {
-                    institutions.add(it.getValue(GpaInstitution::class.java)!!)
+                    semesters.add(it.getValue(GpaSemester::class.java)!!)
                 }
-                LogHelper.i(TAG, "Institution Size: ${institutions.size}")
-                institutions.sortBy { it.order }
-                instituteProcessAndUpdate()
+                LogHelper.i(TAG, "Number of Semesters in Institution: ${semesters.size}")
+                semesters.sortBy { it.order }
+                semestersProcessAndUpdate()
             }
         })
     }
@@ -110,24 +131,28 @@ class GpaCalcInstitutionListFragment : Fragment() {
         }
     }
 
-    private var retry = false
-
-    private fun instituteProcessAndUpdate() {
+    private fun semestersProcessAndUpdate() {
         val list: ArrayList<DualLineString> = ArrayList()
-        val scoring = callback?.getScoreMap()!!
-        if (scoring.isEmpty() && !retry) {
-            retry = true
-            callback?.updateScoreMap(object: GpaCalcCallback { override fun onCallback() { instituteProcessAndUpdate() } })
+        semesters.forEach {
+            list.add(DualLineString(it.name, (if (scoreObject?.type == "count") "Score" else "GPA") + ": TODO"))
         }
-        institutions.forEach {
-            val type = if (scoring.containsKey(it.type)) scoring[it.type]?.name else "Unknown"
-            list.add(DualLineString("${it.name} (${it.shortName})", "Scoring Mode: $type"))
-        }
+        updateActionBar()
         adapter.update(list)
         adapter.notifyDataSetChanged()
     }
 
+    private fun updateActionBar() {
+        var subtitle: String? = null
+        var title: String? = null
+        if (scoreObject != null) {
+            subtitle = if (scoreObject!!.type == "count") "Score" else "GPA"
+            subtitle += ": TODO"
+        }
+        if (selectedInstitution != null) title = "${selectedInstitution!!.name} (${selectedInstitution!!.shortName})"
+        callback?.updateActionBar(title, subtitle)
+    }
+
     companion object {
-        private const val TAG = "GpaCalcInstitutionList"
+        private const val TAG = "GpaCalcSemesterList"
     }
 }
