+++ app/src/main/res/layout/activity_gpa_calculator_add_semester.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:orientation="vertical"
-    tools:context=".Modules.gpaCalculator.GpaCalculatorAddInstitutionActivity">
+    tools:context=".Modules.gpaCalculator.AddSemesterActivity">
 
     <com.google.android.material.textfield.TextInputLayout
         android:layout_width="match_parent"
