+++ app/src/main/res/menu/menu_gpa_calculator_main.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.GpaCalculatorMainActivity">
+    tools:context="com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.MainViewActivity">
     <item android:id="@+id/logout" android:title="Logout"
         android:orderInCategory="50" app:showAsAction="never" />
     <item android:id="@+id/score_system" android:orderInCategory="49"
